#9686 preliminary check in
